@@ -2,17 +2,18 @@
   "name": "well-manager",
   "version": "1.3.16",
   "private": true,
-  "homepage": "/id3x/client/",
+  "type": "module",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "start": "vite",
+    "build": "tsc && vite build",
+    "lint": "eslint . --ext ts,tsx --max-warnings 7",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@cyntler/react-doc-viewer": "1.13.0",
     "@progress/kendo-file-saver": "1.1.1",
     "@progress/kendo-licensing": "1.3.0",
+    "@progress/kendo-react-animation": "5.9.0",
     "@progress/kendo-react-buttons": "5.9.0",
     "@progress/kendo-react-common": "5.6.0",
     "@progress/kendo-react-dateinputs": "5.9.0",
@@ -33,45 +34,40 @@
     "cldr-core": "41.0.0",
     "cldr-dates-full": "41.0.0",
     "cldr-numbers-full": "41.0.0",
-    "flexlayout-react": "0.6.9",
+    "flexlayout-react": "0.7.9",
     "i18next": "21.8.11",
     "jspdf": "2.5.1",
     "lodash": "4.17.21",
     "pako": "2.1.0",
     "parse-color": "1.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "18.2.0",
+    "react-dom": "18.2.0",
     "react-i18next": "11.17.2",
-    "react-redux": "7.2.6",
+    "react-redux": "8.0.5",
     "react-router-dom": "6.3.0",
-    "react-scripts": "5.0.1",
     "recharts": "2.8.0",
     "recharts-to-png": "2.2.1",
     "redux": "4.2.0",
     "redux-thunk": "2.4.1",
+    "stream": "0.0.2",
     "xml-js": "1.6.11"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-private-property-in-object": "7.21.11",
-    "@types/jest": "27.5.2",
     "@types/lodash": "4.14.184",
-    "@types/node": "20.2.1",
     "@types/pako": "2.0.0",
     "@types/parse-color": "1.0.1",
-    "@types/react": "18.0.14",
-    "@types/react-dom": "18.0.5",
+    "@types/react": "18.2.21",
+    "@types/react-dom": "18.2.7",
     "@types/react-router-dom": "5.3.3",
-    "buffer": "6.0.3",
-    "cross-env": "7.0.3",
+    "@typescript-eslint/eslint-plugin": "6.5.0",
+    "@typescript-eslint/parser": "6.5.0",
+    "@vitejs/plugin-react": "4.0.4",
+    "eslint": "8.48.0",
+    "eslint-plugin-react-hooks": "4.6.0",
+    "eslint-plugin-react-refresh": "0.4.3",
     "sass": "1.66.1",
-    "stream": "0.0.2",
-    "typescript": "4.9.5"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
+    "typescript": "5.2.2",
+    "vite": "4.4.9"
   },
   "browserslist": {
     "production": [
