@@ -1,8 +1,7 @@
-import { useSelector, useDispatch } from 'react-redux';
+import { useSelector, useDispatch, compareObjects } from 'shared/lib';
 import { ParameterList, updateParamDeep } from 'entities/parameters';
 import { channelDictSelector, getExternalChannels } from 'entities/channels';
 import { stringToTableCell } from 'entities/parameters/lib/table-row';
-import { compareObjects } from 'shared/lib';
 
 
 interface GlobalParamListProps {
